+++ migrations/20250623101955_create_user_balances_table.sql
@@ -1,8 +1,8 @@
 -- +goose Up
 CREATE TABLE user_balances (
 	user_id INTEGER PRIMARY KEY,
-	current_balance NUMERIC NOT NULL DEFAULT 0,
-	total_withdrawn NUMERIC NOT NULL DEFAULT 0
+	current_balance DOUBLE PRECISION NOT NULL DEFAULT 0,
+	total_withdrawn DOUBLE PRECISION NOT NULL DEFAULT 0
 );
 
 -- +goose Down
